# tmp/svc-music-openapi.json
{"openapi":"3.1.0","info":{"title":"svc-music","version":"dev"},"paths":{"/api/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/health/ready":{"get":{"tags":["health"],"summary":"Ready","operationId":"ready_api_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/music/catalog":{"get":{"tags":["music_catalog"],"summary":"Get Music Catalog","description":"MVP catalog for Music Studio UI.\n\nYou can later move these into DB tables and add /admin sync endpoints,\nsimilar to svc-audio catalog.","operationId":"get_music_catalog_api_music_catalog_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicCatalogResponse"}}}}}}},"/api/music/projects":{"post":{"tags":["music-projects"],"summary":"Create Project","operationId":"create_project_api_music_projects_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMusicProjectIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMusicProjectOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/projects/{project_id}/voice-reference":{"post":{"tags":["music-projects"],"summary":"Upload Voice Reference For Project","operationId":"upload_voice_reference_for_project_api_music_projects__project_id__voice_reference_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_voice_reference_for_project_api_music_projects__project_id__voice_reference_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoiceReferenceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["music-projects"],"summary":"Get Voice Reference For Project","operationId":"get_voice_reference_for_project_api_music_projects__project_id__voice_reference_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoiceReferenceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/projects/{project_id}":{"get":{"tags":["music-projects"],"summary":"Get Project","operationId":"get_project_api_music_projects__project_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicProjectOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["music-projects"],"summary":"Update Project","operationId":"update_project_api_music_projects__project_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMusicProjectIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicProjectOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/projects/{project_id}/performers":{"get":{"tags":["music-projects"],"summary":"List Performers","operationId":"list_performers_api_music_projects__project_id__performers_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MusicPerformerOut"},"title":"Response List Performers Api Music Projects  Project Id  Performers Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["music-projects"],"summary":"Upsert Performer","operationId":"upsert_performer_api_music_projects__project_id__performers_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpsertMusicPerformerIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MusicPerformerOut"},"title":"Response Upsert Performer Api Music Projects  Project Id  Performers Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/projects/{project_id}/lyrics":{"get":{"tags":["music-projects"],"summary":"Get Lyrics","operationId":"get_lyrics_api_music_projects__project_id__lyrics_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicLyricsOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["music-projects"],"summary":"Upsert Lyrics","operationId":"upsert_lyrics_api_music_projects__project_id__lyrics_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpsertMusicLyricsIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicLyricsOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/projects/{project_id}/generate":{"post":{"tags":["music-jobs"],"summary":"Generate","operationId":"generate_api_music_projects__project_id__generate_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Project Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateMusicIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateMusicOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/jobs/{job_id}/status":{"get":{"tags":["music-jobs"],"summary":"Status","operationId":"status_api_music_jobs__job_id__status_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Job Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MusicJobStatusOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/jobs/{job_id}/publish":{"post":{"tags":["music-jobs"],"summary":"Publish","operationId":"publish_api_music_jobs__job_id__publish_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Job Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublishMusicIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublishMusicOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/assets/upload":{"post":{"tags":["music_assets"],"summary":"Upload Music Asset","operationId":"upload_music_asset_api_music_assets_upload_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"kind","in":"query","required":true,"schema":{"type":"string","title":"Kind"}},{"name":"project_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Project Id"}},{"name":"job_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Job Id"}},{"name":"duration_ms","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Ms"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_music_asset_api_music_assets_upload_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssetUploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/assets/{artifact_id}":{"get":{"tags":["music_assets"],"summary":"Get Music Asset","operationId":"get_music_asset_api_music_assets__artifact_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"artifact_id","in":"path","required":true,"schema":{"type":"string","title":"Artifact Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssetGetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/support/sessions/upsert":{"post":{"tags":["support"],"summary":"Upsert Session","operationId":"upsert_session_api_music_support_sessions_upsert_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupportSessionUpsertIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupportSessionOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/support/events":{"post":{"tags":["support"],"summary":"Append Event","operationId":"append_event_api_music_support_events_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupportEventIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupportEventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/support/admin/events":{"post":{"tags":["support"],"summary":"Admin Append Event","description":"Append an admin/support event to the session stream.\n\nInvariants:\n  - support_events.user_id is NOT NULL (legacy)\n  - admin events must be tied to a user context\n  - we derive impersonated_user_id from support_sessions.user_id","operationId":"admin_append_event_api_music_support_admin_events_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminEventIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupportEventOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/support/admin/events/query":{"post":{"tags":["support"],"summary":"Admin Query Events","operationId":"admin_query_events_api_music_support_admin_events_query_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminQueryIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/music/support/admin/sessions/{session_id}/verify-chain":{"get":{"tags":["support"],"summary":"Admin Verify Chain","operationId":"admin_verify_chain_api_music_support_admin_sessions__session_id__verify_chain_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Session Id"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AdminEventIn":{"properties":{"session_id":{"type":"string","format":"uuid","title":"Session Id"},"kind":{"type":"string","enum":["snapshot","action","user_message","assistant_message","system"],"title":"Kind","default":"action"},"payload":{"type":"object","title":"Payload"},"request_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Id"}},"type":"object","required":["session_id"],"title":"AdminEventIn","description":"Admin/support staff event.\n\nClient sends:\n  - session_id\n  - kind/payload (+ optional request_id)\n\nServer derives impersonated_user_id from support_sessions.user_id\n(source of truth) to satisfy legacy support_events.user_id NOT NULL."},"AdminQueryIn":{"properties":{"project_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Project Id"},"job_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Job Id"},"user_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"User Id"},"surface":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Surface"},"limit":{"type":"integer","title":"Limit","default":200}},"type":"object","title":"AdminQueryIn"},"AssetGetResponse":{"properties":{"artifact_id":{"type":"string","title":"Artifact Id"},"storage_path":{"type":"string","title":"Storage Path"},"sas_url":{"type":"string","title":"Sas Url"},"content_type":{"type":"string","title":"Content Type"},"bytes":{"type":"integer","title":"Bytes"},"sha256":{"type":"string","title":"Sha256"},"kind":{"type":"string","title":"Kind"},"project_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Project Id"},"job_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Job Id"},"media_asset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Media Asset Id"}},"type":"object","required":["artifact_id","storage_path","sas_url","content_type","bytes","sha256","kind"],"title":"AssetGetResponse"},"AssetUploadResponse":{"properties":{"artifact_id":{"type":"string","title":"Artifact Id"},"media_asset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Media Asset Id"},"storage_path":{"type":"string","title":"Storage Path"},"sas_url":{"type":"string","title":"Sas Url"},"content_type":{"type":"string","title":"Content Type"},"bytes":{"type":"integer","title":"Bytes"},"sha256":{"type":"string","title":"Sha256"}},"type":"object","required":["artifact_id","storage_path","sas_url","content_type","bytes","sha256"],"title":"AssetUploadResponse"},"Body_upload_music_asset_api_music_assets_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_music_asset_api_music_assets_upload_post"},"Body_upload_voice_reference_for_project_api_music_projects__project_id__voice_reference_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_voice_reference_for_project_api_music_projects__project_id__voice_reference_post"},"CameraEdit":{"type":"string","enum":["smooth","beat_cut","aggressive"],"title":"CameraEdit"},"CreateMusicProjectIn":{"properties":{"title":{"type":"string","maxLength":200,"title":"Title","default":"Untitled Music Video"},"mode":{"allOf":[{"$ref":"#/components/schemas/MusicProjectMode"}],"default":"autopilot"},"duet_layout":{"allOf":[{"$ref":"#/components/schemas/DuetLayout"}],"default":"split_screen"},"language_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language Hint"},"scene_pack_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scene Pack Id"},"camera_edit":{"allOf":[{"$ref":"#/components/schemas/CameraEdit"}],"default":"beat_cut"},"band_pack":{"items":{"type":"string"},"type":"array","title":"Band Pack"}},"type":"object","title":"CreateMusicProjectIn"},"CreateMusicProjectOut":{"properties":{"project_id":{"type":"string","format":"uuid","title":"Project Id"},"status":{"allOf":[{"$ref":"#/components/schemas/MusicProjectStatus"}],"default":"draft"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["project_id","created_at","updated_at"],"title":"CreateMusicProjectOut"},"DuetLayout":{"type":"string","enum":["split_screen","alternating","same_stage"],"title":"DuetLayout"},"GenerateMusicIn":{"properties":{"seed":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seed"},"quality":{"type":"string","enum":["draft","standard","pro"],"title":"Quality","default":"standard"},"outputs":{"items":{"$ref":"#/components/schemas/MusicTrackType"},"type":"array","title":"Outputs"},"provider_hints":{"type":"object","title":"Provider Hints"},"uploaded_audio_asset_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Uploaded Audio Asset Id"},"uploaded_audio_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uploaded Audio Url"},"track_prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Track Prompt"},"genre_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Genre Hint"},"vibe_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vibe Hint"},"lyrics_source":{"anyOf":[{"type":"string","enum":["generate","upload","none"]},{"type":"null"}],"title":"Lyrics Source"},"lyrics_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lyrics Text"},"lyrics_language_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lyrics Language Hint"}},"type":"object","title":"GenerateMusicIn","description":"Works for all project modes:\n  autopilot: system generates track + (default) lyrics\n  co_create: system generates track; lyrics can be uploaded or generated (default generate)\n  byo:       uses uploaded audio; lyrics can be none/upload/generate"},"GenerateMusicOut":{"properties":{"job_id":{"type":"string","format":"uuid","title":"Job Id"},"status":{"$ref":"#/components/schemas/MusicJobStatus"}},"type":"object","required":["job_id","status"],"title":"GenerateMusicOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MusicCatalogResponse":{"properties":{"band_packs":{"items":{"type":"object"},"type":"array","title":"Band Packs"},"scene_packs":{"items":{"type":"object"},"type":"array","title":"Scene Packs"},"camera_edits":{"items":{"type":"object"},"type":"array","title":"Camera Edits"},"duet_layouts":{"items":{"type":"object"},"type":"array","title":"Duet Layouts"},"modes":{"items":{"type":"object"},"type":"array","title":"Modes"}},"type":"object","required":["band_packs","scene_packs","camera_edits","duet_layouts","modes"],"title":"MusicCatalogResponse"},"MusicJobStage":{"type":"string","enum":["intent","creative_brief","lyrics","arrangement","provider_route","generate_audio","align_lyrics","generate_performer_videos","compose_video","qc","publish"],"title":"MusicJobStage"},"MusicJobStatus":{"type":"string","enum":["queued","running","succeeded","failed"],"title":"MusicJobStatus"},"MusicJobStatusOut":{"properties":{"job_id":{"type":"string","format":"uuid","title":"Job Id"},"project_id":{"type":"string","format":"uuid","title":"Project Id"},"status":{"$ref":"#/components/schemas/MusicJobStatus"},"stage":{"anyOf":[{"$ref":"#/components/schemas/MusicJobStage"},{"type":"null"}]},"progress":{"type":"number","title":"Progress","default":0.0},"tracks":{"items":{"$ref":"#/components/schemas/TrackItem"},"type":"array","title":"Tracks"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["job_id","project_id","status"],"title":"MusicJobStatusOut"},"MusicLyricsOut":{"properties":{"project_id":{"type":"string","format":"uuid","title":"Project Id"},"lyrics_text":{"type":"string","title":"Lyrics Text"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["project_id","lyrics_text","created_at","updated_at"],"title":"MusicLyricsOut"},"MusicPerformerOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"project_id":{"type":"string","format":"uuid","title":"Project Id"},"role":{"$ref":"#/components/schemas/MusicPerformerRole"},"image_asset_id":{"type":"string","format":"uuid","title":"Image Asset Id"},"voice_mode":{"$ref":"#/components/schemas/VoiceMode"},"user_is_owner":{"type":"boolean","title":"User Is Owner"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"},"image_content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Content Type"}},"type":"object","required":["id","project_id","role","image_asset_id","voice_mode","user_is_owner","created_at"],"title":"MusicPerformerOut"},"MusicPerformerRole":{"type":"string","enum":["lead","harmony","rap","backing","adlib","narration"],"title":"MusicPerformerRole"},"MusicProjectMode":{"type":"string","enum":["autopilot","co_create","byo"],"title":"MusicProjectMode"},"MusicProjectOut":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"title":{"type":"string","title":"Title"},"mode":{"$ref":"#/components/schemas/MusicProjectMode"},"duet_layout":{"$ref":"#/components/schemas/DuetLayout"},"language_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language Hint"},"scene_pack_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scene Pack Id"},"camera_edit":{"$ref":"#/components/schemas/CameraEdit"},"band_pack":{"items":{"type":"string"},"type":"array","title":"Band Pack"},"status":{"$ref":"#/components/schemas/MusicProjectStatus"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","title","mode","duet_layout","camera_edit","status","created_at","updated_at"],"title":"MusicProjectOut"},"MusicProjectStatus":{"type":"string","enum":["draft","planning","ready","rendering","succeeded","failed"],"title":"MusicProjectStatus"},"MusicTrackType":{"type":"string","enum":["instrumental","vocals","full_mix","stems_zip","lyrics_json","timed_lyrics_json","cover_art"],"title":"MusicTrackType"},"PublishMusicIn":{"properties":{"target":{"type":"string","enum":["viewer","fusion"],"title":"Target","default":"fusion"},"consent":{"type":"object","title":"Consent"}},"type":"object","title":"PublishMusicIn"},"PublishMusicOut":{"properties":{"status":{"type":"string","title":"Status"},"video_job_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Video Job Id"},"fusion_payload":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Fusion Payload"}},"type":"object","required":["status"],"title":"PublishMusicOut"},"SupportEventIn":{"properties":{"session_id":{"type":"string","format":"uuid","title":"Session Id"},"kind":{"type":"string","enum":["snapshot","action","user_message","assistant_message","system"],"title":"Kind"},"payload":{"type":"object","title":"Payload"},"request_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Id"}},"type":"object","required":["session_id","kind"],"title":"SupportEventIn"},"SupportEventOut":{"properties":{"event_id":{"type":"string","format":"uuid","title":"Event Id"},"created_at":{"type":"string","title":"Created At"},"event_hash_hex":{"type":"string","title":"Event Hash Hex"},"prev_hash_hex":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Prev Hash Hex"}},"type":"object","required":["event_id","created_at","event_hash_hex","prev_hash_hex"],"title":"SupportEventOut"},"SupportSessionOut":{"properties":{"session_id":{"type":"string","format":"uuid","title":"Session Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"project_id":{"type":"string","format":"uuid","title":"Project Id"},"job_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Job Id"},"surface":{"type":"string","title":"Surface"},"status":{"type":"string","title":"Status"}},"type":"object","required":["session_id","user_id","project_id","job_id","surface","status"],"title":"SupportSessionOut"},"SupportSessionUpsertIn":{"properties":{"project_id":{"type":"string","format":"uuid","title":"Project Id"},"job_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Job Id"},"surface":{"type":"string","minLength":1,"title":"Surface","default":"music_studio"}},"type":"object","required":["project_id"],"title":"SupportSessionUpsertIn"},"TrackItem":{"properties":{"track_type":{"$ref":"#/components/schemas/MusicTrackType"},"artifact_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Artifact Id"},"media_asset_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Media Asset Id"},"duration_ms":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Ms"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Type"}},"type":"object","required":["track_type"],"title":"TrackItem"},"UpdateMusicProjectIn":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"mode":{"anyOf":[{"$ref":"#/components/schemas/MusicProjectMode"},{"type":"null"}]},"duet_layout":{"anyOf":[{"$ref":"#/components/schemas/DuetLayout"},{"type":"null"}]},"language_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language Hint"},"scene_pack_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scene Pack Id"},"camera_edit":{"anyOf":[{"$ref":"#/components/schemas/CameraEdit"},{"type":"null"}]},"band_pack":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Band Pack"},"status":{"anyOf":[{"$ref":"#/components/schemas/MusicProjectStatus"},{"type":"null"}]}},"type":"object","title":"UpdateMusicProjectIn"},"UpsertMusicLyricsIn":{"properties":{"lyrics_text":{"type":"string","minLength":1,"title":"Lyrics Text"}},"type":"object","required":["lyrics_text"],"title":"UpsertMusicLyricsIn"},"UpsertMusicPerformerIn":{"properties":{"role":{"allOf":[{"$ref":"#/components/schemas/MusicPerformerRole"}],"default":"lead"},"image_asset_id":{"type":"string","format":"uuid","title":"Image Asset Id"},"voice_mode":{"allOf":[{"$ref":"#/components/schemas/VoiceMode"}],"default":"uploaded"},"user_is_owner":{"type":"boolean","title":"User Is Owner","default":false}},"type":"object","required":["image_asset_id"],"title":"UpsertMusicPerformerIn"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VoiceMode":{"type":"string","enum":["uploaded","generated","none"],"title":"VoiceMode"},"VoiceReferenceOut":{"properties":{"project_id":{"type":"string","format":"uuid","title":"Project Id"},"voice_ref_asset_id":{"type":"string","format":"uuid","title":"Voice Ref Asset Id"},"content_type":{"type":"string","title":"Content Type"},"bytes":{"type":"integer","minimum":0.0,"title":"Bytes"},"storage_ref":{"type":"string","title":"Storage Ref"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"}},"type":"object","required":["project_id","voice_ref_asset_id","content_type","bytes","storage_ref"],"title":"VoiceReferenceOut"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}